events_once: finish review hardening

Copilot-Session: 36ab3370-7de7-421b-937c-ba78f88517c9
