+++ packages/events_once/src/core/mod.rs
@@ -5,6 +5,7 @@ mod local_endpoints_nice;
 mod local_receiver;
 mod local_refs;
 mod local_sender;
+mod release_guard;
 mod state;
 mod sync;
 mod sync_embedded;
@@ -20,6 +21,7 @@ pub use local_endpoints_nice::*;
 pub(crate) use local_receiver::*;
 pub(crate) use local_refs::*;
 pub(crate) use local_sender::*;
+pub(crate) use release_guard::*;
 pub(crate) use state::*;
 pub use sync::*;
 pub use sync_embedded::*;
