+++ packages/events_once/src/core/sync.rs
@@ -5,7 +5,7 @@ use std::cell::UnsafeCell;
 use std::fmt;
 use std::hint::spin_loop;
 use std::marker::PhantomPinned;
-use std::mem::{MaybeUninit, offset_of};
+use std::mem::MaybeUninit;
 use std::panic::{RefUnwindSafe, UnwindSafe};
 use std::pin::Pin;
 use std::ptr::NonNull;
@@ -93,9 +93,9 @@ pub struct Event<T> {
 //   invokes any callback. An unwind can therefore never strand the event in `EVENT_SIGNALING`
 //   and leave another endpoint waiting for a transition that will not come.
 // * User-controlled code that can unwind - waker clones, wakes, waker drops and payload drops -
-//   therefore always runs where state and storage agree, so a caught panic leaves an event that
-//   the remaining endpoint can still complete or clean up. The worst outcome is that the event
-//   storage is leaked, which is safe. The unwinding regression tests below exercise this.
+//   therefore always runs where state and storage agree. Cleanup owners guard storage release
+//   across destructor callbacks, so a caught panic leaves the pool or allocation reusable. The
+//   unwinding regression tests below exercise this.
 impl<T: Send + 'static> UnwindSafe for Event<T> {}
 impl<T: Send + 'static> RefUnwindSafe for Event<T> {}
 
@@ -109,39 +109,14 @@ where
     /// wire up the sender and receiver after doing the initialization. An event
     /// without a sender and receiver is invalid.
     pub(crate) fn new_in_inner(place: &mut UnsafeCell<MaybeUninit<Self>>) {
-        // The key here is that we can skip initializing the MaybeUninit fields because
-        // they start uninitialized by design and the UnsafeCell wrapper is transparent,
-        // only affecting accesses and not the contents.
-        let base_ptr = place.get_mut().as_mut_ptr();
-
-        // SAFETY: We are making a pointer to a known field at a compiler-guaranteed offset.
-        let state_ptr = unsafe { base_ptr.byte_add(offset_of!(Self, state)) }.cast::<AtomicU8>();
-
-        // SAFETY: `place` is an exclusive reference to storage laid out and aligned as
-        // `MaybeUninit<Self>`, `get_mut()` preserves that exclusive access and `as_mut_ptr()`
-        // yields its base address, so `offset_of!` selects the in-bounds `state` field with the
-        // alignment of that field. The event lifecycle guarantees that no endpoint of an earlier
-        // event in this storage is still active, so nothing else accesses these bytes. `write`
-        // initializes the field without reading or dropping the previous bytes, which is what
-        // lets us leave `awaiter` and `value` uninitialized.
-        unsafe {
-            state_ptr.write(AtomicU8::new(EVENT_BOUND));
-        }
-
-        #[cfg(debug_assertions)]
-        {
-            // SAFETY: We are making a pointer to a known field at a compiler-guaranteed offset.
-            let backtrace_ptr = unsafe { base_ptr.byte_add(offset_of!(Self, backtrace)) }
-                .cast::<Mutex<Option<BacktraceType>>>();
-
-            // SAFETY: Same as the `state` write above - the exclusive `MaybeUninit<Self>` storage
-            // is aligned writable storage for `Self` that no active event aliases, `offset_of!`
-            // selects the in-bounds `backtrace` field with that field's alignment, and `write`
-            // initializes it without reading or dropping the previous bytes.
-            unsafe {
-                backtrace_ptr.write(Mutex::new(None));
-            }
-        }
+        place.get_mut().write(Self {
+            state: AtomicU8::new(EVENT_BOUND),
+            awaiter: UnsafeCell::new(MaybeUninit::uninit()),
+            value: UnsafeCell::new(MaybeUninit::uninit()),
+            #[cfg(debug_assertions)]
+            backtrace: Mutex::new(None),
+            _requires_pinning: PhantomPinned,
+        });
     }
 
     #[must_use]
@@ -311,9 +286,10 @@ where
 
     /// Sets the value of the event and notifies the receiver's awaiter, if there is one.
     ///
-    /// Returns `Err` if the receiver has already disconnected and we must clean up the event now.
+    /// Returns the undelivered value if the receiver has already disconnected and the caller must
+    /// clean up the event now.
     #[inline]
-    pub(crate) fn set(event_cell: &UnsafeCell<Self>, value: T) -> Result<(), Disconnected> {
+    pub(crate) fn set(event_cell: &UnsafeCell<Self>, value: T) -> Result<(), T> {
         // SAFETY: We only ever create shared references to the event, so no aliasing conflicts.
         // The event lives until both sender and receiver are dropped or inert, so we know it must
         // still exist because something was able to call this method.
@@ -403,15 +379,17 @@ where
             }
             EVENT_DISCONNECTED => {
                 // The receiver has already been dropped, so we need to clean up the event.
-                // We have to first drop the value that we inserted into the event, though.
+                // Move the undelivered value back out before running its destructor so the state
+                // and storage agree throughout that callback.
 
-                // SAFETY: The receiver is gone - there is nobody else who might be touching
-                // the event anymore, we are essentially in a single-threaded mode now.
-                // We also just inserted the value, so it must still be there because we never
-                // entered a state where the receiver had the permission to extract the value.
-                unsafe {
-                    event.destroy_value();
-                }
+                // The receiver is gone, so nobody else can touch the event. The value was just
+                // inserted and the event never entered a state that allowed the receiver to
+                // extract it.
+                let value = event.poll_set();
+
+                event
+                    .state
+                    .store(EVENT_DISCONNECTED, atomic::Ordering::Relaxed);
 
                 // Before it is safe to destroy the event, we need to synchronize with whatever
                 // writes the receiver may have done into its state (e.g. it may have removed
@@ -419,7 +397,7 @@ where
                 atomic::fence(atomic::Ordering::Acquire);
 
                 // The sender (the caller) needs to clean up the event.
-                Err(Disconnected)
+                Err(value)
             }
             // Defensive: state machine guarantees this is unreachable.
             _ => {
@@ -841,8 +819,13 @@ where
     /// Returns `Ok(Some(value))` if the sender sender has already sent a value.
     /// Returns `Err` if the sender has already disconnected without sending a value.
     /// In both of these cases, the receiver must clean up the event now.
+    ///
+    /// The optional waker is returned instead of dropped here so the endpoint core can install a
+    /// storage-release guard before invoking its destructor.
     #[inline]
-    pub(crate) fn final_poll(event_cell: &UnsafeCell<Self>) -> Result<Option<T>, Disconnected> {
+    pub(crate) fn final_poll(
+        event_cell: &UnsafeCell<Self>,
+    ) -> (Result<Option<T>, Disconnected>, Option<Waker>) {
         // SAFETY: We only ever create shared references to the event, so no aliasing conflicts.
         // The event lives until both sender and receiver are dropped or inert, so we know it must
         // still exist because something was able to call this method.
@@ -857,13 +840,11 @@ where
             .expect(NEVER_POISONED)
             .replace(capture_backtrace());
 
-        // The receiver (who is calling this) may still own the waker if the waker has not
-        // been used. If this is the case, we need to destroy the waker before proceeding.
-        // This is only the case if we are in the AWAITING state - in all other states, we
-        // either do not have a waker or the sender will destroy it.
-        // We implement this check by attempting to revert ourselves back to the BOUND state.
-        // If successful, we know we were in AWAITING and can destroy the waker.
-        if event
+        // The receiver may still own a stored waker. Reverting AWAITING -> BOUND acquires
+        // exclusive access to that cell so we can extract the waker without running its
+        // destructor. The terminal transition below is published before the endpoint core
+        // invokes that destructor.
+        let awaiter = if event
             .state
             .compare_exchange(
                 EVENT_AWAITING,
@@ -876,10 +857,10 @@ where
             // SAFETY: The `awaiter` is guaranteed to be present because we just came from
             // `EVENT_AWAITING`. The sender will only touch the awaiter in `EVENT_SIGNALING`,
             // which never entered. Therefore, we are the only one who can touch the awaiter now.
-            unsafe {
-                event.destroy_awaiter();
-            }
-        }
+            Some(unsafe { event.take_awaiter() })
+        } else {
+            None
+        };
 
         // We must transition the event into DISCONNECTED, but we cannot do so while the
         // sender is in the middle of a SIGNALING transition — writing DISCONNECTED via swap
@@ -916,7 +897,7 @@ where
             }
         };
 
-        match previous_state {
+        let result = match previous_state {
             EVENT_BOUND => {
                 // The sender had not yet set any value. It will clean up the event later.
                 Ok(None)
@@ -948,7 +929,25 @@ where
                     type_name::<Self>()
                 );
             }
-        }
+        };
+
+        (result, awaiter)
+    }
+
+    /// Takes the waker registered in `awaiter`, leaving that cell uninitialized.
+    ///
+    /// # Safety
+    ///
+    /// The caller must have acquired the synchronization block for `awaiter` and `awaiter` must
+    /// hold an initialized waker.
+    unsafe fn take_awaiter(&self) -> Waker {
+        // SAFETY: Forwarding guarantees from the caller.
+        let awaiter_cell_maybe = unsafe { self.awaiter.get().as_mut() };
+        // SAFETY: UnsafeCell pointer is never null.
+        let awaiter_cell = unsafe { awaiter_cell_maybe.unwrap_unchecked() };
+
+        // SAFETY: Forwarding guarantees from the caller.
+        unsafe { awaiter_cell.assume_init_read() }
     }
 
     /// Drops the waker registered in `awaiter`, leaving that cell uninitialized.
@@ -966,24 +965,6 @@ where
         // SAFETY: Forwarding guarantees from the caller.
         unsafe {
             awaiter_cell.assume_init_drop();
-        }
-    }
-
-    /// Drops the payload stored in `value`, leaving that cell uninitialized.
-    ///
-    /// # Safety
-    ///
-    /// The caller must have acquired the synchronization block for `value` and `value` must hold
-    /// an initialized payload.
-    unsafe fn destroy_value(&self) {
-        // SAFETY: Forwarding guarantees from the caller.
-        let value_cell_maybe = unsafe { self.value.get().as_mut() };
-        // SAFETY: UnsafeCell pointer is never null.
-        let value_cell = unsafe { value_cell_maybe.unwrap_unchecked() };
-
-        // SAFETY: Forwarding guarantees from the caller.
-        unsafe {
-            value_cell.assume_init_drop();
         }
     }
 }
@@ -1092,1829 +1073,4 @@ mod test_hooks {
 
 #[cfg(test)]
 #[cfg_attr(coverage_nightly, coverage(off))]
-mod tests {
-    use std::cell::RefCell;
-    use std::panic::{AssertUnwindSafe, RefUnwindSafe, UnwindSafe, catch_unwind, resume_unwind};
-    use std::rc::Rc;
-    use std::sync::Barrier;
-    use std::task::Poll;
-    use std::{mem, task, thread};
-
-    use futures::executor::block_on;
-    use static_assertions::assert_impl_all;
-    use testing::{
-        DropOnWakerRelease, ReentrantWakerData, assert_panics_with, clone_action_waker,
-        clone_action_waker_panicking_on_clone_release, drop_waker, with_watchdog,
-    };
-
-    use super::test_hooks::{
-        HOOK_PARTICIPANT, HOOK_POLL_AWAITING_PRE_CAS, HOOK_POLL_BOUND_PRE_CAS,
-        HOOK_SERIALIZATION_MUTEX, HOOK_SET_IN_SIGNALING, HookFn,
-    };
-    use super::*;
-    use crate::IntoValueError;
-
-    assert_impl_all!(Event<u32>: Send, Sync, UnwindSafe, RefUnwindSafe);
-
-    // `Box<dyn Send>` and `&'static mut u32` satisfy the payload bounds of the public API
-    // (`Send + 'static`) while lacking auto traits of their own: the trait object is neither
-    // `Sync` nor unwind-safe in either form, and the exclusive reference is not `UnwindSafe`.
-    // These assertions therefore only pass if the event supplies thread safety and unwind safety
-    // through its own state machine instead of inheriting them structurally from the payload,
-    // which is the contract the endpoints and containers depend on. Preserving `Send` for a trait
-    // object payload is also a regression test for #142.
-    assert_impl_all!(Event<Box<dyn Send>>: Send, Sync, UnwindSafe, RefUnwindSafe);
-    assert_impl_all!(Event<&'static mut u32>: Send, Sync, UnwindSafe, RefUnwindSafe);
-
-    #[test]
-    fn boxed_send_receive() {
-        let (sender, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(42);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-    }
-
-    #[test]
-    fn boxed_send_receive_unit() {
-        let (sender, receiver) = Event::<()>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(());
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok(()))));
-    }
-
-    #[test]
-    fn boxed_send_receive_u128() {
-        let (sender, receiver) = Event::<u128>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(42);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-    }
-
-    #[test]
-    fn boxed_send_receive_array() {
-        let (sender, receiver) = Event::<[u128; 4]>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        sender.send([42, 43, 44, 45]);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok([42, 43, 44, 45]))));
-    }
-
-    #[test]
-    fn boxed_receive_send_receive() {
-        let (sender, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        sender.send(42);
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-    }
-
-    #[test]
-    fn boxed_drop_send() {
-        let (sender, _) = Event::<i32>::boxed();
-
-        sender.send(42);
-    }
-
-    #[test]
-    fn boxed_drop_receive() {
-        let (_, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-    }
-
-    #[test]
-    fn boxed_receive_drop_receive() {
-        let (sender, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(sender);
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-    }
-
-    #[test]
-    fn boxed_receive_drop_send() {
-        let (sender, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(receiver);
-
-        sender.send(42);
-    }
-
-    #[test]
-    fn boxed_receive_drop_drop_receiver_first() {
-        let (sender, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(receiver);
-        drop(sender);
-    }
-
-    #[test]
-    fn boxed_receive_drop_drop_sender_first() {
-        let (sender, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(sender);
-        drop(receiver);
-    }
-
-    #[test]
-    fn boxed_drop_drop_receiver_first() {
-        let (sender, receiver) = Event::<i32>::boxed();
-
-        drop(receiver);
-        drop(sender);
-    }
-
-    #[test]
-    fn boxed_drop_drop_sender_first() {
-        let (sender, receiver) = Event::<i32>::boxed();
-
-        drop(sender);
-        drop(receiver);
-    }
-
-    #[test]
-    fn boxed_is_ready() {
-        let (sender, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        assert!(!receiver.is_ready());
-
-        sender.send(42);
-
-        assert!(receiver.is_ready());
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-    }
-
-    #[test]
-    fn boxed_drop_is_ready() {
-        let (sender, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        assert!(!receiver.is_ready());
-
-        drop(sender);
-
-        assert!(receiver.is_ready());
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-    }
-
-    #[test]
-    fn boxed_into_value() {
-        let (sender, receiver) = Event::<i32>::boxed();
-
-        let Err(IntoValueError::Pending(receiver)) = receiver.into_value() else {
-            panic!("expected no value yet");
-        };
-
-        sender.send(42);
-
-        assert!(matches!(receiver.into_value(), Ok(42)));
-    }
-
-    #[test]
-    fn boxed_drop_into_value() {
-        let (sender, receiver) = Event::<i32>::boxed();
-
-        drop(sender);
-
-        assert!(matches!(
-            receiver.into_value(),
-            Err(IntoValueError::Disconnected)
-        ));
-    }
-
-    #[test]
-    #[should_panic]
-    fn boxed_panic_poll_after_completion() {
-        let (sender, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(42);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        assert!(matches!(
-            receiver.as_mut().poll(&mut cx),
-            Poll::Ready(Ok(42))
-        ));
-
-        // Should panic - invalid to access receiver after it completes.
-        _ = receiver.as_mut().poll(&mut cx);
-    }
-
-    #[test]
-    #[should_panic]
-    fn boxed_panic_is_ready_after_completion() {
-        let (sender, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(42);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        assert!(matches!(
-            receiver.as_mut().poll(&mut cx),
-            Poll::Ready(Ok(42))
-        ));
-
-        // Should panic - invalid to access receiver after it completes.
-        _ = receiver.is_ready();
-    }
-
-    #[test]
-    fn placed_send_receive() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(42);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-    }
-
-    #[test]
-    fn placed_receive_send_receive() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        sender.send(42);
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-    }
-
-    #[test]
-    fn placed_drop_send() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, _) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-        sender.send(42);
-    }
-
-    #[test]
-    fn placed_drop_receive() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (_, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-    }
-
-    #[test]
-    fn placed_receive_drop_receive() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(sender);
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-    }
-
-    #[test]
-    fn placed_receive_drop_send() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(receiver);
-
-        sender.send(42);
-    }
-
-    #[test]
-    fn placed_receive_drop_drop_receiver_first() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(receiver);
-        drop(sender);
-    }
-
-    #[test]
-    fn placed_receive_drop_drop_sender_first() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(sender);
-        drop(receiver);
-    }
-
-    #[test]
-    fn placed_drop_drop_receiver_first() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-        drop(receiver);
-        drop(sender);
-    }
-
-    #[test]
-    fn placed_drop_drop_sender_first() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-        drop(sender);
-        drop(receiver);
-    }
-
-    #[test]
-    fn placed_is_ready() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-        let mut receiver = Box::pin(receiver);
-
-        assert!(!receiver.is_ready());
-
-        sender.send(42);
-
-        assert!(receiver.is_ready());
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-    }
-
-    #[test]
-    fn placed_drop_is_ready() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-        let mut receiver = Box::pin(receiver);
-
-        assert!(!receiver.is_ready());
-
-        drop(sender);
-
-        assert!(receiver.is_ready());
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-    }
-
-    #[test]
-    fn placed_into_value() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-        let Err(IntoValueError::Pending(receiver)) = receiver.into_value() else {
-            panic!("expected no value yet");
-        };
-
-        sender.send(42);
-
-        assert!(matches!(receiver.into_value(), Ok(42)));
-    }
-
-    #[test]
-    fn placed_drop_into_value() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-        drop(sender);
-
-        assert!(matches!(
-            receiver.into_value(),
-            Err(IntoValueError::Disconnected)
-        ));
-    }
-
-    #[test]
-    #[should_panic]
-    fn placed_panic_poll_after_completion() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(42);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        assert!(matches!(
-            receiver.as_mut().poll(&mut cx),
-            Poll::Ready(Ok(42))
-        ));
-
-        // Should panic - invalid to access receiver after it completes.
-        _ = receiver.as_mut().poll(&mut cx);
-    }
-
-    #[test]
-    #[should_panic]
-    fn placed_panic_is_ready_after_completion() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(42);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        assert!(matches!(
-            receiver.as_mut().poll(&mut cx),
-            Poll::Ready(Ok(42))
-        ));
-
-        // Should panic - invalid to access receiver after it completes.
-        _ = receiver.is_ready();
-    }
-
-    #[test]
-    fn boxed_send_receive_mt() {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-
-            thread::spawn(move || {
-                sender.send(42);
-            })
-            .join()
-            .unwrap();
-
-            thread::spawn(move || {
-                let mut receiver = Box::pin(receiver);
-                let mut cx = task::Context::from_waker(Waker::noop());
-
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-            })
-            .join()
-            .unwrap();
-        });
-    }
-
-    #[test]
-    fn boxed_receive_send_receive_mt() {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-
-            let first_poll_completed = Arc::new(Barrier::new(2));
-            let first_poll_completed_clone = Arc::clone(&first_poll_completed);
-
-            let send_thread = thread::spawn(move || {
-                first_poll_completed.wait();
-
-                sender.send(42);
-            });
-
-            let receive_thread = thread::spawn(move || {
-                let mut receiver = Box::pin(receiver);
-                let mut cx = task::Context::from_waker(Waker::noop());
-
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Pending));
-
-                first_poll_completed_clone.wait();
-
-                // We do not know how many polls this will take, so we switch into real async.
-                block_on(async {
-                    let result = &mut receiver.await;
-                    assert!(matches!(result, Ok(42)));
-                });
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    #[test]
-    fn boxed_send_receive_unbiased_mt() {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-
-            let receive_thread = thread::spawn(move || {
-                block_on(async {
-                    let result = &mut receiver.await;
-                    assert!(matches!(result, Ok(42)));
-                });
-            });
-
-            let send_thread = thread::spawn(move || {
-                sender.send(42);
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    #[test]
-    fn boxed_drop_receive_unbiased_mt() {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-
-            let receive_thread = thread::spawn(move || {
-                block_on(async {
-                    let result = &mut receiver.await;
-                    assert!(matches!(result, Err(Disconnected)));
-                });
-            });
-
-            let send_thread = thread::spawn(move || {
-                drop(sender);
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    #[test]
-    fn boxed_drop_send_unbiased_mt() {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-
-            let receive_thread = thread::spawn(move || {
-                drop(receiver);
-            });
-
-            let send_thread = thread::spawn(move || {
-                sender.send(42);
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    #[test]
-    fn placed_send_receive_mt() {
-        with_watchdog(|| {
-            let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-            // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-            // its storage stays allocated, writable and at a stable address until the end of
-            // the test, which outlives both endpoints.
-            let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-            thread::spawn(move || {
-                sender.send(42);
-            })
-            .join()
-            .unwrap();
-
-            thread::spawn(move || {
-                let mut receiver = Box::pin(receiver);
-                let mut cx = task::Context::from_waker(Waker::noop());
-
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-            })
-            .join()
-            .unwrap();
-        });
-    }
-
-    // The repeated thread creation and storage reuse in this test is native-run stress: every
-    // iteration exercises the same lifecycle, so interpreting all of them under each Miri
-    // scheduler seed costs a great deal without reaching new behavior. Miri still covers a single
-    // placed lifecycle via `placed_send_receive_mt` and the other placed tests.
-    #[test]
-    #[cfg_attr(miri, ignore)] // Repeated thread creation and reuse loop is for native runs only.
-    fn placed_send_receive_reused_mt() {
-        with_watchdog(|| {
-            const ITERATIONS: usize = 123;
-
-            let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-
-            for _ in 0..ITERATIONS {
-                // SAFETY: `place` is box-pinned, so its storage stays allocated, writable and at
-                // a stable address until the end of the test, which outlives every endpoint. The
-                // endpoints of the previous iteration are both dropped before we get here, so no
-                // other event is using this storage.
-                let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-                thread::spawn(move || {
-                    sender.send(42);
-                })
-                .join()
-                .unwrap();
-
-                thread::spawn(move || {
-                    let mut receiver = Box::pin(receiver);
-                    let mut cx = task::Context::from_waker(Waker::noop());
-
-                    let poll_result = receiver.as_mut().poll(&mut cx);
-                    assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-                })
-                .join()
-                .unwrap();
-            }
-        });
-    }
-
-    #[test]
-    fn placed_receive_send_receive_mt() {
-        with_watchdog(|| {
-            let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-            // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-            // its storage stays allocated, writable and at a stable address until the end of
-            // the test, which outlives both endpoints.
-            let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-            let first_poll_completed = Arc::new(Barrier::new(2));
-            let first_poll_completed_clone = Arc::clone(&first_poll_completed);
-
-            let send_thread = thread::spawn(move || {
-                first_poll_completed.wait();
-
-                sender.send(42);
-            });
-
-            let receive_thread = thread::spawn(move || {
-                let mut receiver = Box::pin(receiver);
-                let mut cx = task::Context::from_waker(Waker::noop());
-
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Pending));
-
-                first_poll_completed_clone.wait();
-
-                // We do not know how many polls this will take, so we switch into real async.
-                block_on(async {
-                    let result = &mut receiver.await;
-                    assert!(matches!(result, Ok(42)));
-                });
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    #[test]
-    fn placed_send_receive_unbiased_mt() {
-        with_watchdog(|| {
-            let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-            // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-            // its storage stays allocated, writable and at a stable address until the end of
-            // the test, which outlives both endpoints.
-            let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-            let receive_thread = thread::spawn(move || {
-                block_on(async {
-                    let result = &mut receiver.await;
-                    assert!(matches!(result, Ok(42)));
-                });
-            });
-
-            let send_thread = thread::spawn(move || {
-                sender.send(42);
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    #[test]
-    fn placed_drop_receive_unbiased_mt() {
-        with_watchdog(|| {
-            let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-            // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-            // its storage stays allocated, writable and at a stable address until the end of
-            // the test, which outlives both endpoints.
-            let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-            let receive_thread = thread::spawn(move || {
-                block_on(async {
-                    let result = &mut receiver.await;
-                    assert!(matches!(result, Err(Disconnected)));
-                });
-            });
-
-            let send_thread = thread::spawn(move || {
-                drop(sender);
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    #[test]
-    fn placed_drop_send_unbiased_mt() {
-        with_watchdog(|| {
-            let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-            // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-            // its storage stays allocated, writable and at a stable address until the end of
-            // the test, which outlives both endpoints.
-            let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-            let receive_thread = thread::spawn(move || {
-                drop(receiver);
-            });
-
-            let send_thread = thread::spawn(move || {
-                sender.send(42);
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    /// Returns a shared reference to the event placed in `place`, so a test can inspect
-    /// diagnostic state that the endpoints do not expose.
-    ///
-    /// # Safety
-    ///
-    /// An event must have been placed in `place` by [`Event::placed`] and that storage must not
-    /// have been reused for another event since. No exclusive reference to the event storage may
-    /// exist for the lifetime of the returned reference.
-    #[cfg(debug_assertions)]
-    unsafe fn placed_event(place: &EmbeddedEvent<i32>) -> &Event<i32> {
-        // SAFETY: The caller guarantees that an event was placed in this storage, so the pointer
-        // is non-null and aligned for `Event<i32>`, and the storage lives as long as the borrow
-        // of the container. Every access to a placed event - here and in the endpoints - goes
-        // through shared references, and the caller guarantees that no exclusive reference to the
-        // storage exists, so this shared reference cannot conflict with another borrow.
-        let cell = unsafe { place.inner.get().as_ref() }.expect("UnsafeCell pointer is never null");
-
-        // SAFETY: `Event::placed` initialized the event in this storage. Releasing an event only
-        // clears its diagnostic state and never deinitializes those bytes, so the event remains
-        // initialized for as long as the storage exists.
-        unsafe { cell.assume_init_ref() }
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn awaiter_backtrace_no_awaiter() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let _endpoints = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-        // SAFETY: We placed an event in this storage just above and did not reuse the storage.
-        // The endpoints hold only raw pointers and the placement borrow has ended, so there is no
-        // exclusive reference to the storage.
-        let backtrace = unsafe { placed_event(&place) }.awaiter_backtrace();
-
-        assert!(backtrace.is_none());
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn awaiter_backtrace_with_awaiter() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (_sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-        let mut receiver = Box::pin(receiver);
-        _ = receiver.as_mut().poll(&mut cx);
-
-        // SAFETY: We placed an event in this storage above and did not reuse the storage. The
-        // endpoints hold only raw pointers and the placement borrow has ended, so there is no
-        // exclusive reference to the storage.
-        let backtrace = unsafe { placed_event(&place) }.awaiter_backtrace();
-
-        assert!(backtrace.is_some());
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn awaiter_backtrace_after_sender_drop() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-        let mut receiver = Box::pin(receiver);
-        _ = receiver.as_mut().poll(&mut cx);
-
-        drop(sender);
-
-        // SAFETY: We placed an event in this storage above and did not reuse the storage. The
-        // receiver is still an endpoint of it and holds only a raw pointer, so there is no
-        // exclusive reference to the storage.
-        let backtrace = unsafe { placed_event(&place) }.awaiter_backtrace();
-
-        assert!(backtrace.is_some());
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn awaiter_backtrace_after_receiver_drop() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-        // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-        // its storage stays allocated, writable and at a stable address until the end of
-        // the test, which outlives both endpoints.
-        let (_sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-        let mut receiver = Box::pin(receiver);
-        _ = receiver.as_mut().poll(&mut cx);
-
-        drop(receiver);
-
-        // SAFETY: We placed an event in this storage above and did not reuse the storage. The
-        // sender is still an endpoint of it and holds only a raw pointer, so there is no
-        // exclusive reference to the storage.
-        let backtrace = unsafe { placed_event(&place) }.awaiter_backtrace();
-
-        assert!(backtrace.is_some());
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn awaiter_backtrace_outlives_event() {
-        let backtrace = {
-            let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-            // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-            // its storage stays allocated, writable and at a stable address until the end of
-            // the test, which outlives both endpoints.
-            let (_sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-            let mut cx = task::Context::from_waker(Waker::noop());
-            let mut receiver = Box::pin(receiver);
-            _ = receiver.as_mut().poll(&mut cx);
-
-            // SAFETY: We placed an event in this storage above and did not reuse the storage.
-            // The endpoints hold only raw pointers and the placement borrow has ended, so there
-            // is no exclusive reference to the storage.
-            unsafe { placed_event(&place) }
-                .awaiter_backtrace()
-                .expect("the event has been awaited")
-        };
-
-        // The event storage is gone but the snapshot remains readable.
-        _ = backtrace.status();
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn released_event_releases_backtrace() {
-        let mut place = Box::pin(EmbeddedEvent::<i32>::new());
-
-        {
-            // SAFETY: `place` is a fresh container that holds no other event, box-pinned so
-            // its storage stays allocated, writable and at a stable address until the end of
-            // the test, which outlives both endpoints.
-            let (_sender, receiver) = unsafe { Event::<i32>::placed(place.as_mut()) };
-
-            let mut cx = task::Context::from_waker(Waker::noop());
-            let mut receiver = Box::pin(receiver);
-            _ = receiver.as_mut().poll(&mut cx);
-        }
-
-        // The event has been released but its storage is still ours to inspect. Releasing an
-        // event releases its backtrace, because the storage may be reused without dropping it.
-        //
-        // SAFETY: We placed an event in this storage above and did not reuse the storage.
-        // Releasing the event left it initialized, both endpoints are gone and nothing holds an
-        // exclusive reference to the storage.
-        let event = unsafe { placed_event(&place) };
-
-        assert!(event.awaiter_backtrace().is_none());
-    }
-
-    /// Installs a hook closure and runs the test body while holding the
-    /// serialization mutex. The hook is always removed on exit.
-    fn with_hook(hook: &Mutex<Option<Arc<HookFn>>>, closure: Arc<HookFn>, body: impl FnOnce()) {
-        let guard = HOOK_SERIALIZATION_MUTEX.lock().unwrap();
-        *hook.lock().unwrap() = Some(closure);
-
-        // We catch panics from the test body so that we can clean up the hook and
-        // drop the serialization guard while not panicking, preventing mutex poisoning.
-        let result = catch_unwind(AssertUnwindSafe(body));
-
-        *hook.lock().unwrap() = None;
-        drop(guard);
-
-        if let Err(payload) = result {
-            resume_unwind(payload);
-        }
-    }
-
-    struct BarrierHook {
-        /// Signaled when the hook fires, indicating that the hooked code
-        /// has reached the synchronization point.
-        entered: Arc<Barrier>,
-
-        /// Waited on before the hook returns, giving the test thread a
-        /// window to perform a racing operation.
-        proceed: Arc<Barrier>,
-
-        /// The closure to install as a hook.
-        hook: Arc<HookFn>,
-    }
-
-    /// Creates a two-barrier hook closure that pauses execution at a
-    /// synchronization point, giving the test thread a window to perform
-    /// a racing operation.
-    fn barrier_hook() -> BarrierHook {
-        let entered = Arc::new(Barrier::new(2));
-        let proceed = Arc::new(Barrier::new(2));
-        let e = Arc::clone(&entered);
-        let p = Arc::clone(&proceed);
-        let hook: Arc<HookFn> = Arc::new(move || {
-            e.wait();
-            p.wait();
-        });
-        BarrierHook {
-            entered,
-            proceed,
-            hook,
-        }
-    }
-
-    #[test]
-    fn boxed_poll_bound_races_sender_disconnect() {
-        with_watchdog(|| {
-            let BarrierHook {
-                entered,
-                proceed,
-                hook,
-            } = barrier_hook();
-            with_hook(&HOOK_POLL_BOUND_PRE_CAS, hook, || {
-                let (sender, receiver) = Event::<i32>::boxed();
-
-                // Receiver polls on a separate thread. It will write the
-                // waker and then pause at the hook, before the CAS.
-                let receive_thread = thread::spawn(move || {
-                    HOOK_PARTICIPANT.set(true);
-                    let mut receiver = Box::pin(receiver);
-                    let mut cx = task::Context::from_waker(Waker::noop());
-                    receiver.as_mut().poll(&mut cx)
-                });
-
-                // Wait for the hook to fire (receiver wrote waker).
-                entered.wait();
-
-                // Drop the sender while the receiver is paused. This
-                // transitions BOUND -> SIGNALING -> DISCONNECTED.
-                drop(sender);
-
-                // Release the receiver so its CAS(BOUND->AWAITING) fails
-                // with DISCONNECTED.
-                proceed.wait();
-
-                let poll_result = receive_thread.join().unwrap();
-                assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-            });
-        });
-    }
-
-    #[test]
-    fn boxed_poll_awaiting_races_sender_set() {
-        with_watchdog(|| {
-            let BarrierHook {
-                entered,
-                proceed,
-                hook,
-            } = barrier_hook();
-            with_hook(&HOOK_POLL_AWAITING_PRE_CAS, hook, || {
-                let (sender, receiver) = Event::<i32>::boxed();
-                let mut receiver = Box::pin(receiver);
-
-                // First poll transitions BOUND -> AWAITING.
-                let mut cx = task::Context::from_waker(Waker::noop());
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Pending));
-
-                // Second poll on a separate thread enters poll_awaiting
-                // and pauses at the hook before the CAS.
-                let receive_thread = thread::spawn(move || {
-                    HOOK_PARTICIPANT.set(true);
-                    let mut cx = task::Context::from_waker(Waker::noop());
-                    receiver.as_mut().poll(&mut cx)
-                });
-
-                // Wait for the hook to fire.
-                entered.wait();
-
-                // Send value while the receiver is paused. This
-                // transitions AWAITING -> SIGNALING -> SET.
-                sender.send(42);
-
-                // Release the receiver so its CAS(AWAITING->BOUND) fails
-                // with SET.
-                proceed.wait();
-
-                let poll_result = receive_thread.join().unwrap();
-                assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-            });
-        });
-    }
-
-    #[test]
-    fn boxed_poll_awaiting_races_sender_disconnect() {
-        with_watchdog(|| {
-            let BarrierHook {
-                entered,
-                proceed,
-                hook,
-            } = barrier_hook();
-            with_hook(&HOOK_POLL_AWAITING_PRE_CAS, hook, || {
-                let (sender, receiver) = Event::<i32>::boxed();
-                let mut receiver = Box::pin(receiver);
-
-                // First poll transitions BOUND -> AWAITING.
-                let mut cx = task::Context::from_waker(Waker::noop());
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Pending));
-
-                // Second poll on a separate thread enters poll_awaiting
-                // and pauses at the hook before the CAS.
-                let receive_thread = thread::spawn(move || {
-                    HOOK_PARTICIPANT.set(true);
-                    let mut cx = task::Context::from_waker(Waker::noop());
-                    receiver.as_mut().poll(&mut cx)
-                });
-
-                // Wait for the hook to fire.
-                entered.wait();
-
-                // Drop sender while the receiver is paused. This
-                // transitions AWAITING -> SIGNALING -> DISCONNECTED.
-                drop(sender);
-
-                // Release the receiver so its CAS(AWAITING->BOUND) fails
-                // with DISCONNECTED.
-                proceed.wait();
-
-                let poll_result = receive_thread.join().unwrap();
-                assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-            });
-        });
-    }
-
-    // This test verifies that `final_poll` correctly handles the case where the sender is
-    // mid-SIGNALING when the receiver is dropped. The fix uses a CAS loop in `final_poll`
-    // that spins on SIGNALING, so the receiver waits for the sender to finish before writing
-    // DISCONNECTED. The receiver-drop must happen on a separate thread because `final_poll`
-    // will spin until the sender completes, and the sender is blocked on the hook barrier.
-    #[test]
-    // The receiver's `final_poll` spins while the event is in SIGNALING, and here the sender only
-    // leaves SIGNALING once the test releases its barrier. Miri's interpreter makes such a
-    // cross-thread spin extremely slow, so this scenario is reserved for native runs.
-    #[cfg_attr(miri, ignore)]
-    fn boxed_final_poll_races_sender_signaling() {
-        with_watchdog(|| {
-            let BarrierHook {
-                entered,
-                proceed,
-                hook,
-            } = barrier_hook();
-            with_hook(&HOOK_SET_IN_SIGNALING, hook, || {
-                let (sender, receiver) = Event::<i32>::boxed();
-                let mut receiver = Box::pin(receiver);
-
-                // First poll transitions BOUND -> AWAITING.
-                let mut cx = task::Context::from_waker(Waker::noop());
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Pending));
-
-                // Sender sends on a separate thread. The fetch_add
-                // transitions AWAITING -> SIGNALING. The sender then
-                // pauses at the hook in the SIGNALING state.
-                let send_thread = thread::spawn(move || {
-                    HOOK_PARTICIPANT.set(true);
-                    sender.send(42);
-                });
-
-                // Wait for the hook to fire (sender is in SIGNALING).
-                entered.wait();
-
-                // Drop the receiver on a separate thread. final_poll
-                // will spin on SIGNALING until the sender completes
-                // its transition, so we cannot block this thread — we
-                // need it to release the sender via proceed.wait().
-                let drop_thread = thread::spawn(move || {
-                    drop(receiver);
-                });
-
-                // Release the sender so it can complete its transition
-                // from SIGNALING -> SET. This unblocks final_poll's
-                // spin, which then sees SET and reads the value.
-                proceed.wait();
-
-                send_thread.join().unwrap();
-                drop_thread.join().unwrap();
-            });
-        });
-    }
-
-    // Regression test for issue #462: `is_ready()` must not report readiness while the
-    // sender is mid-`set()` in the transient SIGNALING state, because `into_value()` still
-    // reports `Pending` there. The two must agree. `EVENT_SIGNALING` is not a terminal
-    // state, so it must count as "not ready".
-    #[test]
-    fn boxed_is_ready_false_while_sender_signaling() {
-        with_watchdog(|| {
-            let BarrierHook {
-                entered,
-                proceed,
-                hook,
-            } = barrier_hook();
-            with_hook(&HOOK_SET_IN_SIGNALING, hook, || {
-                let (sender, receiver) = Event::<i32>::boxed();
-                let mut receiver = Box::pin(receiver);
-
-                // First poll transitions BOUND -> AWAITING so that the sender's
-                // `set()` takes the AWAITING path that pauses at the hook.
-                let mut cx = task::Context::from_waker(Waker::noop());
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Pending));
-
-                // Sender sends on a separate thread. The `fetch_add` in `set` transitions
-                // AWAITING -> SIGNALING, then the sender pauses at the hook while still in the
-                // SIGNALING state. (The sender-disconnect path reaches SIGNALING via `swap`
-                // instead; see `sender_dropped_without_set`.)
-                let send_thread = thread::spawn(move || {
-                    HOOK_PARTICIPANT.set(true);
-                    sender.send(42);
-                });
-
-                // Wait for the hook to fire (sender is parked in SIGNALING).
-                entered.wait();
-
-                // The value is not yet retrievable in SIGNALING, so readiness must be
-                // false here, consistent with `into_value()` reporting `Pending`.
-                assert!(!receiver.is_ready());
-
-                // Release the sender so it completes SIGNALING -> SET.
-                proceed.wait();
-                send_thread.join().unwrap();
-
-                // Now in the terminal SET state, readiness agrees with retrieval.
-                assert!(receiver.is_ready());
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-            });
-        });
-    }
-
-    // Synchronous extraction has a concurrency boundary that `Future::poll` does not cover: it
-    // classifies the observed state and, for a terminal state, finalizes the event and may
-    // release its storage while the sender is still acting. The transient SIGNALING state is the
-    // interesting classification, because there the sender owns the event fields but has not
-    // published an outcome yet - an extraction must report pending and hand back a usable
-    // receiver. The hook that parks a sender in SIGNALING is private to this module, which is why
-    // this test lives here rather than beside the other `into_value` tests in `sync_receiver.rs`.
-    // Ref: docs/testing.md, "Testing atomic operations and custom synchronization".
-    #[test]
-    fn boxed_into_value_pending_while_sender_signaling() {
-        with_watchdog(|| {
-            let BarrierHook {
-                entered,
-                proceed,
-                hook,
-            } = barrier_hook();
-            with_hook(&HOOK_SET_IN_SIGNALING, hook, || {
-                let (sender, mut receiver) = Event::<i32>::boxed();
-
-                // First poll transitions BOUND -> AWAITING, which is what routes the sender's
-                // `set` through SIGNALING.
-                let mut cx = task::Context::from_waker(Waker::noop());
-                assert!(matches!(
-                    Pin::new(&mut receiver).poll(&mut cx),
-                    Poll::Pending
-                ));
-
-                let send_thread = thread::spawn(move || {
-                    HOOK_PARTICIPANT.set(true);
-                    sender.send(42);
-                });
-
-                // Wait for the hook to fire (sender is parked in SIGNALING, holding the awaiter).
-                entered.wait();
-
-                if receiver.is_ready() {
-                    // Consuming or dropping the receiver now would finalize the event, which
-                    // spins until the sender leaves SIGNALING - and only this thread can release
-                    // the sender. We leak the receiver, release the sender and fail, so that a
-                    // regression is reported instead of hanging when no watchdog is active.
-                    mem::forget(receiver);
-                    proceed.wait();
-                    send_thread.join().unwrap();
-
-                    panic!("SIGNALING must not be reported as a completed event");
-                }
-
-                let extraction = receiver.into_value();
-
-                // Release the sender before asserting, so that a failure cannot leave it parked
-                // in the hook.
-                proceed.wait();
-                send_thread.join().unwrap();
-
-                let Err(IntoValueError::Pending(receiver)) = extraction else {
-                    panic!("the value must not be extractable while the sender is signaling");
-                };
-
-                // The retained receiver is still connected to the event, so it observes the value
-                // that the sender has since published.
-                assert_eq!(
-                    receiver.into_value().expect("the sender published a value"),
-                    42
-                );
-            });
-        });
-    }
-
-    // Regression test for the synchronous reentrancy hazard in
-    // `sender_dropped_without_set`. A waker fired by the sender drop that
-    // synchronously polls the receiver must observe a terminal state
-    // (DISCONNECTED), not the transient SIGNALING state — otherwise the
-    // reentrant poll would spin in `poll_signaling` while the sender is
-    // blocked inside `wake()`, producing a same-thread deadlock.
-    #[test]
-    #[cfg_attr(miri, ignore)] // Custom raw waker is not Miri-compatible.
-    fn boxed_sender_drop_with_reentrant_waker_does_not_deadlock() {
-        type ObservedResult = Poll<Result<i32, Disconnected>>;
-
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-            let receiver_holder: Rc<RefCell<Option<Pin<Box<_>>>>> =
-                Rc::new(RefCell::new(Some(Box::pin(receiver))));
-            let receiver_for_waker = Rc::clone(&receiver_holder);
-
-            let reentrant_observed: Rc<RefCell<Option<ObservedResult>>> =
-                Rc::new(RefCell::new(None));
-            let observed_for_waker = Rc::clone(&reentrant_observed);
-
-            let waker_data = ReentrantWakerData::new(move || {
-                // Synchronously poll the receiver from inside the waker.
-                // With the buggy ordering this would enter `poll_signaling`
-                // and spin while we are still blocked inside `wake()`.
-                let mut holder = receiver_for_waker.borrow_mut();
-                let receiver = holder.as_mut().expect("receiver still held");
-                let noop = Waker::noop();
-                let mut cx = task::Context::from_waker(noop);
-                let result = receiver.as_mut().poll(&mut cx);
-                *observed_for_waker.borrow_mut() = Some(result);
-            });
-            // SAFETY: `waker_data` outlives the waker and the test is single-threaded.
-            let waker = unsafe { waker_data.waker() };
-
-            // First poll transitions BOUND -> AWAITING and stores the
-            // reentrant waker.
-            {
-                let mut holder = receiver_holder.borrow_mut();
-                let receiver = holder.as_mut().expect("receiver still held");
-                let mut cx = task::Context::from_waker(&waker);
-                assert!(matches!(receiver.as_mut().poll(&mut cx), Poll::Pending));
-            }
-
-            // Drop the sender. This calls `sender_dropped_without_set`,
-            // which transitions AWAITING -> SIGNALING, then must
-            // transition to DISCONNECTED before invoking the waker so
-            // that the reentrant poll observes a terminal state.
-            drop(sender);
-
-            assert!(waker_data.was_woken());
-            let observed = reentrant_observed.borrow_mut().take();
-            assert!(
-                matches!(observed, Some(Poll::Ready(Err(Disconnected)))),
-                "reentrant poll should observe DISCONNECTED",
-            );
-
-            // Drop the receiver to release its half of the event.
-            drop(receiver_holder.borrow_mut().take());
-        });
-    }
-
-    // Parity counterpart of the disconnect case above. A waker fired by `send` that
-    // synchronously polls the receiver must likewise observe a terminal state, here SET, and read
-    // out the value.
-    #[test]
-    #[cfg_attr(miri, ignore)] // Custom raw waker is not Miri-compatible.
-    fn boxed_send_with_reentrant_waker_observes_set() {
-        type ObservedResult = Poll<Result<i32, Disconnected>>;
-
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-            let receiver_holder: Rc<RefCell<Option<Pin<Box<_>>>>> =
-                Rc::new(RefCell::new(Some(Box::pin(receiver))));
-            let receiver_for_waker = Rc::clone(&receiver_holder);
-
-            let reentrant_observed: Rc<RefCell<Option<ObservedResult>>> =
-                Rc::new(RefCell::new(None));
-            let observed_for_waker = Rc::clone(&reentrant_observed);
-
-            let waker_data = ReentrantWakerData::new(move || {
-                let mut holder = receiver_for_waker.borrow_mut();
-                let receiver = holder.as_mut().expect("receiver still held");
-                let noop = Waker::noop();
-                let mut cx = task::Context::from_waker(noop);
-                let result = receiver.as_mut().poll(&mut cx);
-                *observed_for_waker.borrow_mut() = Some(result);
-            });
-            // SAFETY: `waker_data` outlives the waker and the test is single-threaded.
-            let waker = unsafe { waker_data.waker() };
-
-            // First poll transitions BOUND -> AWAITING and stores the reentrant waker.
-            {
-                let mut holder = receiver_holder.borrow_mut();
-                let receiver = holder.as_mut().expect("receiver still held");
-                let mut cx = task::Context::from_waker(&waker);
-                assert!(matches!(receiver.as_mut().poll(&mut cx), Poll::Pending));
-            }
-
-            // `set` reaches a terminal state and then invokes the waker, which must observe SET
-            // and consume the value reentrantly.
-            sender.send(42);
-
-            assert!(waker_data.was_woken());
-            let observed = reentrant_observed.borrow_mut().take();
-            assert!(
-                matches!(observed, Some(Poll::Ready(Ok(42)))),
-                "reentrant poll should observe SET and read the value",
-            );
-
-            // The receiver was consumed reentrantly; drop the shell that still owns it.
-            drop(receiver_holder.borrow_mut().take());
-        });
-    }
-
-    // Parity counterparts of the `LocalEvent` reentrancy regression tests in `core/local.rs`.
-    // A wake callback fired while completing or cancelling an event is free to drop the receiver,
-    // which releases the event storage while the operation is still on the stack. Ref:
-    // docs/callback-safety.md. These run under Miri, which is what detects a regression here.
-    #[test]
-    fn boxed_receiver_cancel_with_sender_dropping_waker_preserves_storage() {
-        let (sender, receiver) = Event::<i32>::boxed();
-        let mut receiver = Box::pin(receiver);
-
-        let (data, sender_dropped) = DropOnWakerRelease::new(sender);
-        // SAFETY: The payload is not `Send`, and this test keeps the waker on one thread.
-        let waker = unsafe { drop_waker(data) };
-
-        // First poll transitions BOUND -> AWAITING and stores a clone of the waker in the event.
-        let mut cx = task::Context::from_waker(&waker);
-        assert!(matches!(receiver.as_mut().poll(&mut cx), Poll::Pending));
-
-        // Drop our local waker so only the event's stored clone remains; the sender is still
-        // owned behind that clone.
-        drop(waker);
-        assert!(!sender_dropped.get());
-
-        // Dropping the receiver cancels the wait: `final_poll` drops the stored waker, whose
-        // destructor drops the sender. The event storage must survive until `final_poll` returns.
-        drop(receiver);
-
-        assert!(sender_dropped.get(), "{REENTRANCY_REQUIRED}");
-    }
-
-    /// Owns a receiver on behalf of a waker payload and drops it only once the event has reached
-    /// a terminal state.
-    ///
-    /// The reentrancy tests below release this from inside a sender operation, where dropping the
-    /// receiver re-enters `Event::final_poll`. That call spins while the event is in the
-    /// transient `EVENT_SIGNALING` state, and the sender that would leave that state is the very
-    /// operation blocked in this destructor - so a regression in terminal-state publication would
-    /// hang the test instead of failing it. Checking readiness first turns that regression into a
-    /// deterministic failure, which is what mutation testing needs given that watchdogs are
-    /// disabled there. Ref: docs/testing.md, "Tests must not hang".
-    struct TerminalStateReceiverDrop {
-        receiver: Option<Pin<Box<BoxedReceiver<i32>>>>,
-    }
-
-    impl TerminalStateReceiverDrop {
-        fn new(receiver: BoxedReceiver<i32>) -> Self {
-            Self {
-                receiver: Some(Box::pin(receiver)),
-            }
-        }
-
-        fn receiver_mut(&mut self) -> Pin<&mut BoxedReceiver<i32>> {
-            self.receiver
-                .as_mut()
-                .expect("the receiver is only taken while dropping")
-                .as_mut()
-        }
-    }
-
-    impl Drop for TerminalStateReceiverDrop {
-        fn drop(&mut self) {
-            let receiver = self
-                .receiver
-                .take()
-                .expect("the receiver is only taken while dropping");
-
-            if receiver.is_ready() {
-                drop(receiver);
-                return;
-            }
-
-            // The event has not published a terminal state, so dropping the receiver here would
-            // spin forever waiting for the sender that is blocked in this destructor. We leak the
-            // receiver instead and report the defect.
-            mem::forget(receiver);
-
-            assert!(
-                thread::panicking(),
-                "the sender must publish a terminal state before releasing the stored waker"
-            );
-        }
-    }
-
-    #[test]
-    fn boxed_send_with_reentrant_receiver_drop_releases_storage() {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-
-            let (data, receiver_dropped) =
-                DropOnWakerRelease::new(TerminalStateReceiverDrop::new(receiver));
-            // SAFETY: The payload is not `Send`, and this test keeps the waker on one thread.
-            let waker = unsafe { drop_waker(Arc::clone(&data)) };
-
-            // First poll transitions BOUND -> AWAITING and stores a clone of the waker in the
-            // event.
-            data.with_value(|holder| {
-                let mut cx = task::Context::from_waker(&waker);
-                assert!(matches!(holder.receiver_mut().poll(&mut cx), Poll::Pending));
-            });
-
-            // Leave the event's stored clone as the only reference, so waking it runs the
-            // reentrant drop.
-            drop(waker);
-            drop(data);
-            assert!(!receiver_dropped.get());
-
-            // `set` stores the value, reaches a terminal state and wakes, which drops the
-            // receiver, which consumes the value and frees the event.
-            sender.send(42);
-
-            assert!(receiver_dropped.get(), "{REENTRANCY_REQUIRED}");
-        });
-    }
-
-    #[test]
-    fn boxed_sender_drop_with_reentrant_receiver_drop_releases_storage() {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-
-            let (data, receiver_dropped) =
-                DropOnWakerRelease::new(TerminalStateReceiverDrop::new(receiver));
-            // SAFETY: The payload is not `Send`, and this test keeps the waker on one thread.
-            let waker = unsafe { drop_waker(Arc::clone(&data)) };
-
-            data.with_value(|holder| {
-                let mut cx = task::Context::from_waker(&waker);
-                assert!(matches!(holder.receiver_mut().poll(&mut cx), Poll::Pending));
-            });
-
-            drop(waker);
-            drop(data);
-            assert!(!receiver_dropped.get());
-
-            // `sender_dropped_without_set` reaches DISCONNECTED and wakes, which drops the
-            // receiver, which frees the event.
-            drop(sender);
-
-            assert!(receiver_dropped.get(), "{REENTRANCY_REQUIRED}");
-        });
-    }
-
-    // The re-poll path destroys the registration made by the previous poll before it registers
-    // the replacement. Releasing that registration is user code that may drop the sender, which
-    // completes the event from inside the re-poll: the sender observes EVENT_BOUND and publishes
-    // EVENT_DISCONNECTED without touching the awaiter, so the re-poll must observe that state,
-    // release the replacement it made and report the disconnect.
-    // Ref: docs/callback-safety.md.
-    #[test]
-    fn boxed_repoll_with_reentrant_sender_drop_during_previous_waker_release_observes_disconnected()
-    {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-            let mut receiver = Box::pin(receiver);
-
-            let (data, sender_dropped) = DropOnWakerRelease::new(sender);
-            // SAFETY: The payload is not `Send`, and this test keeps the waker on one thread.
-            let previous_waker = unsafe { drop_waker(data) };
-
-            // First poll transitions BOUND -> AWAITING and registers a clone of `previous_waker`.
-            let mut cx = task::Context::from_waker(&previous_waker);
-            assert!(matches!(receiver.as_mut().poll(&mut cx), Poll::Pending));
-
-            // Release our own reference so that the registration inside the event is the last
-            // one, and destroying it drops the sender it carries.
-            drop(previous_waker);
-            assert!(!sender_dropped.get());
-
-            let (replacement_data, replacement_released) = DropOnWakerRelease::new(());
-            // SAFETY: The payload is not `Send`, and this test keeps the waker on one thread.
-            let replacement_waker = unsafe { drop_waker(Arc::clone(&replacement_data)) };
-
-            let mut cx = task::Context::from_waker(&replacement_waker);
-            let poll_result = receiver.as_mut().poll(&mut cx);
-
-            assert!(sender_dropped.get(), "{REENTRANCY_REQUIRED}");
-            assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-
-            // The event completed, so it must not have kept the replacement registration - once
-            // we release our own references, the payload is dropped.
-            drop(replacement_waker);
-            drop(replacement_data);
-            assert!(
-                replacement_released.get(),
-                "the re-poll must release the replacement registration it made"
-            );
-        });
-    }
-
-    /// Explains a failure to reach the reentrant drop that a regression test exists to exercise.
-    /// Without it the test would pass no matter how the code under test behaved.
-    const REENTRANCY_REQUIRED: &str =
-        "the event must have held a waker clone whose drop reentered the operation under test";
-
-    /// Explains a failure to reach the reentrant clone that a regression test exists to exercise.
-    /// Without it the test would pass no matter how the code under test behaved.
-    const WAKER_CLONE_REQUIRED: &str =
-        "the poll must have cloned the waker it was given, which is what re-enters the sender";
-
-    // Parity counterparts of the `LocalEvent` waker-clone reentrancy regression tests in
-    // `core/local.rs`. Registering an awaiter clones the waker, which is user code that may
-    // operate on the sender endpoint of the same event and move it into a terminal state. The
-    // poll must observe the state the clone left behind. Ref: docs/callback-safety.md.
-    #[test]
-    fn boxed_poll_with_reentrant_send_during_waker_clone_observes_set() {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-            let mut receiver = Box::pin(receiver);
-
-            // SAFETY: The payload is not `Send`, and this test keeps the waker on one thread.
-            let (waker, cloned) = unsafe { clone_action_waker(move || sender.send(42)) };
-
-            let mut cx = task::Context::from_waker(&waker);
-            let poll_result = receiver.as_mut().poll(&mut cx);
-
-            assert!(cloned.get(), "{WAKER_CLONE_REQUIRED}");
-            assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-        });
-    }
-
-    #[test]
-    fn boxed_poll_with_reentrant_sender_drop_during_waker_clone_observes_disconnected() {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-            let mut receiver = Box::pin(receiver);
-
-            // SAFETY: The payload is not `Send`, and this test keeps the waker on one thread.
-            let (waker, cloned) = unsafe { clone_action_waker(move || drop(sender)) };
-
-            let mut cx = task::Context::from_waker(&waker);
-            let poll_result = receiver.as_mut().poll(&mut cx);
-
-            assert!(cloned.get(), "{WAKER_CLONE_REQUIRED}");
-            assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-        });
-    }
-
-    #[test]
-    fn boxed_repoll_with_reentrant_send_during_waker_clone_observes_set() {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-            let mut receiver = Box::pin(receiver);
-
-            // First poll transitions BOUND -> AWAITING and registers a waker for the sender to
-            // take.
-            let mut cx = task::Context::from_waker(Waker::noop());
-            assert!(matches!(receiver.as_mut().poll(&mut cx), Poll::Pending));
-
-            // SAFETY: The payload is not `Send`, and this test keeps the waker on one thread.
-            let (waker, cloned) = unsafe { clone_action_waker(move || sender.send(42)) };
-
-            let mut cx = task::Context::from_waker(&waker);
-            let poll_result = receiver.as_mut().poll(&mut cx);
-
-            assert!(cloned.get(), "{WAKER_CLONE_REQUIRED}");
-            assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-        });
-    }
-
-    #[test]
-    fn boxed_repoll_with_reentrant_sender_drop_during_waker_clone_observes_disconnected() {
-        with_watchdog(|| {
-            let (sender, receiver) = Event::<i32>::boxed();
-            let mut receiver = Box::pin(receiver);
-
-            let mut cx = task::Context::from_waker(Waker::noop());
-            assert!(matches!(receiver.as_mut().poll(&mut cx), Poll::Pending));
-
-            // SAFETY: The payload is not `Send`, and this test keeps the waker on one thread.
-            let (waker, cloned) = unsafe { clone_action_waker(move || drop(sender)) };
-
-            let mut cx = task::Context::from_waker(&waker);
-            let poll_result = receiver.as_mut().poll(&mut cx);
-
-            assert!(cloned.get(), "{WAKER_CLONE_REQUIRED}");
-            assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-        });
-    }
-
-    /// Counts its own drops, so a test can tell a value that was delivered exactly once apart from
-    /// one that was extracted twice or lost.
-    struct DropCounter {
-        drops: Arc<atomic::AtomicUsize>,
-    }
-
-    impl Drop for DropCounter {
-        fn drop(&mut self) {
-            self.drops.fetch_add(1, atomic::Ordering::Relaxed);
-        }
-    }
-
-    // Parity counterpart of the local unwinding test. Releasing the awaiter that a racing send
-    // made useless is user code, so it may unwind. The value must still be inside the event at
-    // that moment, or the receiver's own cleanup reads an extracted cell a second time.
-    // Ref: docs/callback-safety.md.
-    #[test]
-    fn boxed_poll_unwinding_during_waker_clone_release_leaves_value_in_event() {
-        with_watchdog(|| {
-            let drops = Arc::new(atomic::AtomicUsize::new(0));
-
-            let (sender, receiver) = Event::<DropCounter>::boxed();
-            let mut receiver = Box::pin(receiver);
-
-            let value = DropCounter {
-                drops: Arc::clone(&drops),
-            };
-
-            // SAFETY: The payload is not `Send`, and this test keeps the waker on one thread.
-            let (waker, cloned) = unsafe {
-                clone_action_waker_panicking_on_clone_release(move || sender.send(value))
-            };
-
-            let mut cx = task::Context::from_waker(&waker);
-            assert_panics_with(
-                || receiver.as_mut().poll(&mut cx),
-                |message| assert!(message.contains("waker clone release")),
-            );
-
-            assert!(cloned.get(), "{WAKER_CLONE_REQUIRED}");
-
-            if drops.load(atomic::Ordering::Relaxed) != 0 {
-                // The poll extracted the value before running the callback, so the event now
-                // claims a value it no longer holds. Letting the receiver clean up would read
-                // that cell again, so leak the event instead of escalating the failure into
-                // undefined behavior.
-                mem::forget(receiver);
-
-                panic!(
-                    "the value must stay in the event while a callback can still unwind past it"
-                );
-            }
-
-            // The event still owns the value, so the receiver's cleanup delivers exactly one drop.
-            drop(receiver);
-
-            assert_eq!(drops.load(atomic::Ordering::Relaxed), 1);
-        });
-    }
-}
+mod tests;
