+++ packages/events_once/src/lake/raw_sync.rs
@@ -39,7 +39,8 @@ use crate::{NEVER_POISONED, RawEventPool, RawPooledReceiver, RawPooledSender};
 /// where
 ///     T: Send + Debug + 'static,
 /// {
-///     // SAFETY: We promise the lake outlives both the returned endpoints.
+///     // SAFETY: The lake is pinned outside this call, and both endpoints are consumed before
+///     // the function returns, so their backing pool remains alive and stationary.
 ///     let (tx, rx) = unsafe { lake.rent::<T>() };
 ///
 ///     tx.send(payload);
@@ -108,6 +109,7 @@ impl RawEventLake {
     /// Rents an event from the lake, returning its endpoints.
     ///
     /// The event will be returned to the lake when both endpoints are dropped.
+    /// See [`RawPooledReceiver`] for the receiver's callback and reentrancy contract.
     ///
     /// # Safety
     ///
@@ -300,19 +302,19 @@ impl<T: Send + 'static> ErasedPool for PoolWrapper<T> {
 }
 
 #[cfg(test)]
-#[allow(clippy::undocumented_unsafe_blocks, reason = "test code, be concise")]
 #[cfg_attr(coverage_nightly, coverage(off))]
 mod tests {
     use core::task;
     #[cfg(debug_assertions)]
     use std::cell::RefCell;
     use std::panic::{RefUnwindSafe, UnwindSafe};
+    use std::sync::Barrier;
     use std::task::Waker;
+    use std::thread;
 
     use static_assertions::assert_impl_all;
     #[cfg(debug_assertions)]
     use testing::assert_panics_with;
-    #[cfg(debug_assertions)]
     use testing::with_watchdog;
 
     use super::*;
@@ -326,17 +328,104 @@ mod tests {
     );
 
     #[test]
+    fn concurrent_same_type_rentals_are_usable_across_threads() {
+        // The smallest group that guarantees competing map and pool access.
+        const WORKER_COUNT: usize = 2;
+
+        with_watchdog(|| {
+            let lake = Box::pin(RawEventLake::new());
+            let barrier = Barrier::new(WORKER_COUNT);
+
+            thread::scope(|scope| {
+                for _ in 0..WORKER_COUNT {
+                    let lake = lake.as_ref().get_ref();
+                    let barrier = &barrier;
+
+                    scope.spawn(move || {
+                        barrier.wait();
+
+                        // SAFETY: The enclosing scope joins this worker and its sender thread
+                        // before the pinned lake is dropped.
+                        let (sender, receiver) = unsafe { lake.rent::<i32>() };
+                        thread::scope(|scope| {
+                            scope.spawn(move || sender.send(42)).join().unwrap();
+                        });
+
+                        assert_eq!(receiver.into_value().unwrap(), 42);
+                    });
+                }
+            });
+
+            assert!(lake.is_empty());
+        });
+    }
+
+    #[test]
+    fn concurrent_distinct_type_rentals_are_usable_across_threads() {
+        with_watchdog(|| {
+            let lake = Box::pin(RawEventLake::new());
+            let barrier = Barrier::new(2);
+
+            thread::scope(|scope| {
+                scope.spawn({
+                    let lake = lake.as_ref().get_ref();
+                    let barrier = &barrier;
+
+                    move || {
+                        barrier.wait();
+
+                        // SAFETY: The enclosing scope joins this worker and its sender thread
+                        // before the pinned lake is dropped.
+                        let (sender, receiver) = unsafe { lake.rent::<i32>() };
+                        thread::scope(|scope| {
+                            scope.spawn(move || sender.send(42)).join().unwrap();
+                        });
+
+                        assert_eq!(receiver.into_value().unwrap(), 42);
+                    }
+                });
+
+                scope.spawn({
+                    let lake = lake.as_ref().get_ref();
+                    let barrier = &barrier;
+
+                    move || {
+                        barrier.wait();
+
+                        // SAFETY: The enclosing scope joins this worker and its sender thread
+                        // before the pinned lake is dropped.
+                        let (sender, receiver) = unsafe { lake.rent::<String>() };
+                        thread::scope(|scope| {
+                            scope
+                                .spawn(move || sender.send("payload".to_owned()))
+                                .join()
+                                .unwrap();
+                        });
+
+                        assert_eq!(receiver.into_value().unwrap(), "payload");
+                    }
+                });
+            });
+
+            assert!(lake.is_empty());
+        });
+    }
+
+    #[test]
     fn len() {
         let lake = RawEventLake::new();
 
         assert_eq!(lake.len(), 0);
 
+        // SAFETY: The lake remains alive until both returned endpoints are dropped.
         let (sender1, receiver1) = unsafe { lake.rent::<String>() };
         assert_eq!(lake.len(), 1);
 
+        // SAFETY: The lake remains alive until both returned endpoints are dropped.
         let (sender2, receiver2) = unsafe { lake.rent::<i32>() };
         assert_eq!(lake.len(), 2);
 
+        // SAFETY: The lake remains alive until both returned endpoints are dropped.
         let (sender3, receiver3) = unsafe { lake.rent::<String>() };
         assert_eq!(lake.len(), 3);
 
@@ -359,7 +448,9 @@ mod tests {
 
         assert!(lake.is_empty());
 
+        // SAFETY: The lake remains alive until both returned endpoints are dropped.
         let (sender1, receiver1) = unsafe { lake.rent::<String>() };
+        // SAFETY: The lake remains alive until both returned endpoints are dropped.
         let (sender2, receiver2) = unsafe { lake.rent::<i32>() };
 
         assert!(!lake.is_empty());
@@ -389,8 +480,11 @@ mod tests {
         let lake = RawEventLake::new();
 
         // 2 events that are awaited and one that is not.
+        // SAFETY: The lake remains alive until both returned endpoints are dropped.
         let (sender1, receiver1) = unsafe { lake.rent::<String>() };
+        // SAFETY: The lake remains alive until both returned endpoints are dropped.
         let (_sender2, receiver2) = unsafe { lake.rent::<i32>() };
+        // SAFETY: The lake remains alive until both returned endpoints are dropped.
         let (_sender3, _receiver3) = unsafe { lake.rent::<f64>() };
 
         let mut receiver1 = Box::pin(receiver1);
