+++ packages/events_once/src/pool/local.rs
@@ -73,6 +73,7 @@ impl<T: 'static> LocalEventPool<T> {
     /// Rents an event from the pool, returning its endpoints.
     ///
     /// The event will be returned to the pool when both endpoints are dropped.
+    /// See [`PooledLocalReceiver`] for the receiver's callback and reentrancy contract.
     #[inline]
     #[must_use]
     pub fn rent(&self) -> (PooledLocalSender<T>, PooledLocalReceiver<T>) {
@@ -179,6 +180,7 @@ mod tests {
     use static_assertions::{assert_impl_all, assert_not_impl_any};
     #[cfg(debug_assertions)]
     use testing::assert_panics_with;
+    use testing::{PanicsOnDrop, assert_panics, clone_action_waker_panicking_on_clone_release};
 
     use super::*;
     use crate::Disconnected;
@@ -191,6 +193,35 @@ mod tests {
     // The payload satisfies only the bound that the pool's API requires (`'static`) and is
     // neither `UnwindSafe` nor `RefUnwindSafe`, so both traits come from the pool itself.
     assert_impl_all!(LocalEventPool<Rc<RefCell<u32>>>: UnwindSafe, RefUnwindSafe);
+
+    #[test]
+    fn disconnected_send_releases_slot_when_payload_drop_panics() {
+        let pool = LocalEventPool::<PanicsOnDrop>::new();
+        let (sender, receiver) = pool.rent();
+        drop(receiver);
+
+        assert_panics(|| sender.send(PanicsOnDrop));
+        assert!(pool.is_empty());
+    }
+
+    #[test]
+    fn receiver_drop_releases_slot_when_waker_drop_panics() {
+        let pool = LocalEventPool::<i32>::new();
+        let (sender, receiver) = pool.rent();
+        let mut receiver = Box::pin(receiver);
+
+        // SAFETY: The payload is not `Send`, and this test keeps the waker on one thread.
+        let (waker, cloned) = unsafe { clone_action_waker_panicking_on_clone_release(|| {}) };
+        let mut cx = task::Context::from_waker(&waker);
+        assert!(matches!(receiver.as_mut().poll(&mut cx), Poll::Pending));
+        assert!(cloned.get());
+        drop(waker);
+
+        assert_panics(|| drop(receiver));
+        drop(sender);
+
+        assert!(pool.is_empty());
+    }
 
     #[test]
     fn len() {
