+++ packages/events_once/src/pool/raw_sync_endpoints_nice.rs
@@ -47,6 +47,13 @@ impl<T: Send + 'static> fmt::Debug for RawPooledSender<T> {
 /// Awaiting the receiver will yield either the payload of type `T` or a [`Disconnected`] error.
 ///
 /// This kind of endpoint is used for events stored in a raw event pool or event lake.
+///
+/// # Reentrancy
+///
+/// Cloning a waker during polling may synchronously send through or drop the sender. Waking or
+/// dropping a registered waker during completion or cancellation may synchronously poll this
+/// receiver to completion or drop an endpoint. The event publishes the resulting state before
+/// each callback, and catching a callback panic returns its slot to the pool once cleanup completes.
 pub struct RawPooledReceiver<T: Send + 'static> {
     inner: ReceiverCore<RawPooledRef<T>, T>,
 }
