+++ packages/events_once/src/pool/sync.rs
@@ -60,6 +60,7 @@ impl<T: Send + 'static> EventPool<T> {
     /// Rents an event from the pool, returning its endpoints.
     ///
     /// The event will be returned to the pool when both endpoints are dropped.
+    /// See [`PooledReceiver`] for the receiver's callback and reentrancy contract.
     #[inline]
     #[must_use]
     pub fn rent(&self) -> (PooledSender<T>, PooledReceiver<T>) {
@@ -162,671 +163,4 @@ impl<T: Send + 'static> fmt::Debug for EventPoolCore<T> {
 
 #[cfg(test)]
 #[cfg_attr(coverage_nightly, coverage(off))]
-mod tests {
-    #[cfg(debug_assertions)]
-    use std::cell::RefCell;
-    use std::panic::{RefUnwindSafe, UnwindSafe};
-    use std::sync::Barrier;
-    use std::task::{self, Poll, Waker};
-    use std::{iter, thread};
-
-    use futures::executor::block_on;
-    use static_assertions::assert_impl_all;
-    #[cfg(debug_assertions)]
-    use testing::assert_panics_with;
-    use testing::with_watchdog;
-
-    use super::*;
-    use crate::Disconnected;
-    #[cfg(debug_assertions)]
-    use crate::assert_inspect_awaiters_is_reentrant;
-
-    // The payload satisfies only the bound that the pool's API requires (`Send`) and lacks every
-    // trait asserted here, so each of them is supplied by the pool's own synchronization and
-    // storage rather than inherited from the payload. A trait object payload also has to preserve
-    // the thread-safety traits (regression test for #142).
-    assert_impl_all!(EventPool<Box<dyn Send>>: Send, Sync, UnwindSafe, RefUnwindSafe);
-
-    #[test]
-    fn len() {
-        let pool = EventPool::<i32>::new();
-
-        assert_eq!(pool.len(), 0);
-
-        let (sender1, receiver1) = pool.rent();
-        assert_eq!(pool.len(), 1);
-
-        let (sender2, receiver2) = pool.rent();
-        assert_eq!(pool.len(), 2);
-
-        drop(sender1);
-        drop(receiver1);
-        assert_eq!(pool.len(), 1);
-
-        drop(sender2);
-        drop(receiver2);
-        assert_eq!(pool.len(), 0);
-    }
-
-    #[test]
-    fn send_receive() {
-        let pool = EventPool::<i32>::new();
-
-        assert!(pool.is_empty());
-
-        let (sender, receiver) = pool.rent();
-
-        assert!(!pool.is_empty());
-
-        {
-            let mut receiver = Box::pin(receiver);
-
-            sender.send(42);
-
-            let mut cx = task::Context::from_waker(Waker::noop());
-
-            let poll_result = receiver.as_mut().poll(&mut cx);
-            assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-        }
-
-        assert!(pool.is_empty());
-    }
-
-    #[test]
-    fn send_receive_reused() {
-        const ITERATIONS: usize = 32;
-
-        let pool = EventPool::<i32>::new();
-
-        assert!(pool.is_empty());
-
-        for _ in 0..ITERATIONS {
-            let (sender, receiver) = pool.rent();
-            let mut receiver = Box::pin(receiver);
-
-            sender.send(42);
-
-            let mut cx = task::Context::from_waker(Waker::noop());
-
-            let poll_result = receiver.as_mut().poll(&mut cx);
-            assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-        }
-
-        assert!(pool.is_empty());
-    }
-
-    #[test]
-    fn send_receive_reused_batches() {
-        const ITERATIONS: usize = 4;
-        const BATCH_SIZE: usize = 8;
-
-        let pool = EventPool::<i32>::new();
-
-        for _ in 0..ITERATIONS {
-            let endpoints = iter::repeat_with(|| pool.rent())
-                .take(BATCH_SIZE)
-                .collect::<Vec<_>>();
-
-            for (sender, receiver) in endpoints {
-                let mut receiver = Box::pin(receiver);
-
-                sender.send(42);
-
-                let mut cx = task::Context::from_waker(Waker::noop());
-
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-            }
-        }
-    }
-
-    #[test]
-    fn drop_send() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, _) = pool.rent();
-
-        sender.send(42);
-    }
-
-    #[test]
-    fn drop_receive() {
-        let pool = EventPool::<i32>::new();
-
-        let (_, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-    }
-
-    #[test]
-    fn receive_drop_receive() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(sender);
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-    }
-
-    #[test]
-    fn receive_drop_send() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(receiver);
-
-        sender.send(42);
-    }
-
-    #[test]
-    fn receive_drop_drop_receiver_first() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(receiver);
-        drop(sender);
-    }
-
-    #[test]
-    fn receive_drop_drop_sender_first() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Pending));
-
-        drop(sender);
-        drop(receiver);
-    }
-
-    #[test]
-    fn drop_drop_receiver_first() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-
-        drop(receiver);
-        drop(sender);
-    }
-
-    #[test]
-    fn drop_drop_sender_first() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-
-        drop(sender);
-        drop(receiver);
-    }
-
-    #[test]
-    fn is_ready() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        assert!(!receiver.is_ready());
-
-        sender.send(42);
-
-        assert!(receiver.is_ready());
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-    }
-
-    #[test]
-    fn drop_is_ready() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        assert!(!receiver.is_ready());
-
-        drop(sender);
-
-        assert!(receiver.is_ready());
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Err(Disconnected))));
-    }
-
-    #[test]
-    fn into_value() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-
-        let Err(crate::IntoValueError::Pending(receiver)) = receiver.into_value() else {
-            panic!("Expected receiver to not be ready");
-        };
-
-        sender.send(42);
-
-        assert!(matches!(receiver.into_value(), Ok(42)));
-    }
-
-    #[test]
-    #[should_panic]
-    fn panic_poll_after_completion() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(42);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        assert!(matches!(
-            receiver.as_mut().poll(&mut cx),
-            Poll::Ready(Ok(42))
-        ));
-
-        _ = receiver.as_mut().poll(&mut cx);
-    }
-
-    #[test]
-    #[should_panic]
-    fn panic_is_ready_after_completion() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(42);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        assert!(matches!(
-            receiver.as_mut().poll(&mut cx),
-            Poll::Ready(Ok(42))
-        ));
-
-        _ = receiver.is_ready();
-    }
-
-    #[test]
-    fn send_receive_mt() {
-        with_watchdog(|| {
-            let pool = EventPool::<i32>::new();
-
-            let (sender, receiver) = pool.rent();
-
-            thread::spawn(move || {
-                sender.send(42);
-            })
-            .join()
-            .unwrap();
-
-            thread::spawn(move || {
-                let mut receiver = Box::pin(receiver);
-                let mut cx = task::Context::from_waker(Waker::noop());
-
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-            })
-            .join()
-            .unwrap();
-        });
-    }
-
-    #[test]
-    fn receive_send_receive_mt() {
-        with_watchdog(|| {
-            let pool = EventPool::<i32>::new();
-
-            let (sender, receiver) = pool.rent();
-
-            let first_poll_completed = Arc::new(Barrier::new(2));
-            let first_poll_completed_clone = Arc::clone(&first_poll_completed);
-
-            let send_thread = thread::spawn(move || {
-                first_poll_completed.wait();
-
-                sender.send(42);
-            });
-
-            let receive_thread = thread::spawn(move || {
-                let mut receiver = Box::pin(receiver);
-                let mut cx = task::Context::from_waker(Waker::noop());
-
-                let poll_result = receiver.as_mut().poll(&mut cx);
-                assert!(matches!(poll_result, Poll::Pending));
-
-                first_poll_completed_clone.wait();
-
-                // We do not know how many polls this will take, so we switch into real async.
-                block_on(async {
-                    let result = &mut receiver.await;
-                    assert!(matches!(result, Ok(42)));
-                });
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    #[test]
-    fn send_receive_unbiased_mt() {
-        with_watchdog(|| {
-            let pool = EventPool::<i32>::new();
-
-            let (sender, receiver) = pool.rent();
-
-            let receive_thread = thread::spawn(move || {
-                block_on(async {
-                    let result = &mut receiver.await;
-                    assert!(matches!(result, Ok(42)));
-                });
-            });
-
-            let send_thread = thread::spawn(move || {
-                sender.send(42);
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    #[test]
-    fn drop_receive_unbiased_mt() {
-        with_watchdog(|| {
-            let pool = EventPool::<i32>::new();
-
-            let (sender, receiver) = pool.rent();
-
-            let receive_thread = thread::spawn(move || {
-                block_on(async {
-                    let result = &mut receiver.await;
-                    assert!(matches!(result, Err(Disconnected)));
-                });
-            });
-
-            let send_thread = thread::spawn(move || {
-                drop(sender);
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    #[test]
-    fn drop_send_unbiased_mt() {
-        with_watchdog(|| {
-            let pool = EventPool::<i32>::new();
-
-            let (sender, receiver) = pool.rent();
-
-            let receive_thread = thread::spawn(move || {
-                drop(receiver);
-            });
-
-            let send_thread = thread::spawn(move || {
-                sender.send(42);
-            });
-
-            send_thread.join().unwrap();
-            receive_thread.join().unwrap();
-        });
-    }
-
-    #[test]
-    fn drop_pool_send_receive() {
-        let pool = EventPool::<i32>::new();
-
-        let (sender, receiver) = pool.rent();
-
-        drop(pool);
-
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(42);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn inspect_awaiters_inspects_only_awaited() {
-        let pool = EventPool::<i32>::new();
-
-        let (_sender1, receiver1) = pool.rent();
-        let (sender2, receiver2) = pool.rent();
-        let (_sender3, _receiver3) = pool.rent();
-
-        let mut receiver1 = Box::pin(receiver1);
-        let mut receiver2 = Box::pin(receiver2);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-        _ = receiver1.as_mut().poll(&mut cx);
-        _ = receiver2.as_mut().poll(&mut cx);
-
-        let mut inspected_count = 0;
-
-        pool.inspect_awaiters(|_bt| {
-            inspected_count += 1;
-        });
-
-        assert_eq!(inspected_count, 2);
-
-        drop(sender2);
-        drop(receiver2);
-
-        let mut inspected_count = 0;
-
-        pool.inspect_awaiters(|_bt| {
-            inspected_count += 1;
-        });
-
-        assert_eq!(inspected_count, 1);
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn clones_are_equivalent() {
-        let pool1 = EventPool::<i32>::new();
-        let pool2 = pool1.clone();
-
-        let (_sender1, receiver1) = pool1.rent();
-        let (_sender2, receiver2) = pool2.rent();
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let mut receiver1 = Box::pin(receiver1);
-        let mut receiver2 = Box::pin(receiver2);
-
-        _ = receiver1.as_mut().poll(&mut cx);
-        _ = receiver2.as_mut().poll(&mut cx);
-
-        // The inspect_awaiters() logic is sticky, so we can use that to validate.
-        let mut inspected_count = 0;
-
-        pool1.inspect_awaiters(|_bt| {
-            inspected_count += 1;
-        });
-
-        assert_eq!(inspected_count, 2);
-
-        let mut inspected_count = 0;
-
-        pool2.inspect_awaiters(|_bt| {
-            inspected_count += 1;
-        });
-
-        assert_eq!(inspected_count, 2);
-    }
-
-    #[test]
-    fn default_creates_functional_pool() {
-        let pool = EventPool::<i32>::default();
-
-        assert!(pool.is_empty());
-
-        let (sender, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        sender.send(42);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-
-        let poll_result = receiver.as_mut().poll(&mut cx);
-        assert!(matches!(poll_result, Poll::Ready(Ok(42))));
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn inspect_awaiters_propagates_panic_from_closure() {
-        let pool = EventPool::<i32>::new();
-        let (_sender, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-        _ = receiver.as_mut().poll(&mut cx);
-
-        assert_panics_with(
-            || {
-                pool.inspect_awaiters(|_bt| {
-                    panic!("intentional panic to verify pass-through");
-                });
-            },
-            |message| assert!(message.contains("pass-through")),
-        );
-
-        // The pool is still usable, which proves that the panic did not leave any lock behind.
-        assert_eq!(pool.len(), 1);
-
-        let mut inspected_count = 0;
-
-        pool.inspect_awaiters(|_bt| {
-            inspected_count += 1;
-        });
-
-        assert_eq!(inspected_count, 1);
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn inspect_awaiters_closure_may_reenter_pool() {
-        // The callback re-enters the pool, so a regression that calls it under the pool lock
-        // deadlocks on the non-reentrant mutex. The watchdog turns that into a bounded failure.
-        with_watchdog(|| {
-            let pool = EventPool::<i32>::new();
-
-            let (_sender, receiver) = pool.rent();
-            let mut receiver = Box::pin(receiver);
-
-            let mut cx = task::Context::from_waker(Waker::noop());
-            _ = receiver.as_mut().poll(&mut cx);
-
-            assert_inspect_awaiters_is_reentrant(&|f| pool.inspect_awaiters(f), &|| {
-                let (sender, receiver) = pool.rent();
-                drop(sender);
-                drop(receiver);
-            });
-        });
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn inspect_awaiters_tolerates_endpoint_drop_from_closure() {
-        const EVENT_COUNT: usize = 3;
-
-        // Dropping endpoints from the callback returns events to the pool, which takes the pool
-        // lock. The watchdog bounds the deadlock that a callback-under-lock regression causes.
-        with_watchdog(|| {
-            let pool = EventPool::<i32>::new();
-
-            let mut cx = task::Context::from_waker(Waker::noop());
-
-            let mut endpoints = Vec::with_capacity(EVENT_COUNT);
-
-            for _ in 0..EVENT_COUNT {
-                let (sender, receiver) = pool.rent();
-                let mut receiver = Box::pin(receiver);
-                _ = receiver.as_mut().poll(&mut cx);
-                endpoints.push((sender, receiver));
-            }
-
-            // The closure releases the events it is inspecting. The backtraces it receives are
-            // snapshots, so they remain valid and each event is still visited exactly once.
-            let endpoints = RefCell::new(endpoints);
-            let mut inspected_count = 0;
-
-            pool.inspect_awaiters(|_bt| {
-                inspected_count += 1;
-                drop(endpoints.borrow_mut().pop());
-            });
-
-            assert_eq!(inspected_count, EVENT_COUNT);
-            assert!(pool.is_empty());
-        });
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    fn released_event_releases_backtrace() {
-        let pool = EventPool::<i32>::new();
-        let (sender, receiver) = pool.rent();
-        let mut receiver = Box::pin(receiver);
-
-        let mut cx = task::Context::from_waker(Waker::noop());
-        _ = receiver.as_mut().poll(&mut cx);
-
-        // The receiver leaves the event behind for the sender to release.
-        drop(receiver);
-
-        let mut backtraces = pool.awaiter_backtraces();
-        assert_eq!(backtraces.len(), 1);
-
-        let backtrace = backtraces.pop().expect("the event has been awaited");
-        assert_eq!(Arc::strong_count(&backtrace), 2);
-
-        drop(sender);
-
-        // Releasing the event releases its backtrace, leaving the snapshot as the only owner.
-        assert_eq!(Arc::strong_count(&backtrace), 1);
-    }
-}
+mod tests;
