+++ packages/events_once/src/pool/sync/tests.rs
@@ -1,0 +1,29 @@
+#[cfg(debug_assertions)]
+use std::cell::RefCell;
+use std::panic::{RefUnwindSafe, UnwindSafe};
+use std::sync::Barrier;
+use std::task::{self, Poll, Waker};
+use std::{iter, thread};
+
+use futures::executor::block_on;
+use static_assertions::assert_impl_all;
+#[cfg(debug_assertions)]
+use testing::assert_panics_with;
+use testing::{
+    PanicsOnDrop, assert_panics, clone_action_waker_panicking_on_clone_release, with_watchdog,
+};
+
+use super::*;
+use crate::Disconnected;
+#[cfg(debug_assertions)]
+use crate::assert_inspect_awaiters_is_reentrant;
+
+// The payload satisfies only the bound that the pool's API requires (`Send`) and lacks every
+// trait asserted here, so each of them is supplied by the pool's own synchronization and
+// storage rather than inherited from the payload. A trait object payload also has to preserve
+// the thread-safety traits (regression test for #142).
+assert_impl_all!(EventPool<Box<dyn Send>>: Send, Sync, UnwindSafe, RefUnwindSafe);
+
+mod concurrency;
+mod diagnostics;
+mod lifecycle;
